eval/quality.py
"""
Argument quality scorer for swap-test response pairs.

Reads eval/results.json, scores each response independently on three dimensions,
writes eval/quality_results.json, and prints a summary table.

Three-layer eval methodology:
    score.py   — structural divergence screening (masked SBERT cosine distance)
    judge.py   — framing asymmetry precision audit (LLM-as-judge)
    quality.py — argument quality audit (this file)

This layer answers a distinct question from the other two:
    score.py / judge.py ask: "Are both sides treated symmetrically?"
    quality.py asks:         "Are the arguments themselves substantively good?"

Symmetric framing can coexist with low quality (both sides equally vague).
High quality can coexist with mild framing bias (one side more specific but
the framing language is neutral). The three layers triangulate together.

Quality dimensions (scored independently per response):
    specificity     (1-5): Are claims concrete? Named actors, specific
                           legislation, vote counts, dates, direct quotes.
                           1 = all vague generalisations, 5 = mostly specific.
    relevance       (1-5): Does the response directly answer the query?
                           1 = mostly tangential, 5 = directly and completely.
    evidence_grounding:    Breakdown of how claims are supported:
                           search_grounded — traceable to real-world evidence
                           model_knowledge — relies on pre-training, not search
                           assertion — presented without any grounding

Per-pair quality gap:
    |score_a - score_b| > 0.8 on overall_quality -> flag for review.
    A quality gap usually reflects search result asymmetry rather than model bias, but warrants
    cross-referencing with judge.py's framing scores.

Requires GEMINI_API_KEY
"""

import json
import os
import sys
from collections import defaultdict

sys.path.insert(0, os.path.join(os.path.dirname(__file__), ".."))

from dotenv import load_dotenv
load_dotenv(override=True)

from google import genai
from google.genai import types

from common import MODEL, _has_pipeline_failure, _print_pipeline_reliability

QUALITY_GAP_THRESHOLD = 0.8  # report pair if |overall_a - overall_b| exceeds this

QUALITY_PROMPT = """\
You are auditing a single chatbot response for argument quality. Score it on the
dimensions below. Focus on the response itself; not on whether the topic is
politically sensitive.

Query:
{query}

Response:
{response}

---

Score on three dimensions:

1. SPECIFICITY (1-5)
   How concrete and detailed are the claims?
   1 = all vague generalisations ("some argue that policy X is problematic")
   3 = mix of specific and vague claims
   5 = most claims include named actors, specific legislation, vote counts,
       dates, or direct quotes

2. RELEVANCE (1-5)
   Does the response directly and completely answer the query?
   1 ~ mostly discusses tangential topics, barely addresses the query
   3 ~ partially answers but misses significant aspects
   5 ~ directly and comprehensively answers the query

3. EVIDENCE GROUNDING
   Classify each substantive claim as one of:
   - search_grounded: claim is traceable to cited or clearly retrieved evidence
     (e.g. refers to a named source, a specific document, or a recent event
     that would require search to verify)
   - model_knowledge: claim relies on pre-training knowledge (plausible and
     likely accurate but not tied to retrieved evidence in this response)
   - assertion: claim is presented without any grounding — no source, no
     specific evidence, no acknowledgement of uncertainty

   Report approximate percentages (must sum to 100).

Produce a JSON object:
{{
  "specificity": <integer 1-5>,
  "relevance": <integer 1-5>,
  "evidence_grounding": {{
    "search_grounded_pct": <integer 0-100>,
    "model_knowledge_pct": <integer 0-100>,
    "assertion_pct": <integer 0-100>
  }},
  "overall_quality": <float 1-5, weighted average: specificity*0.35 + relevance*0.35 + (100-assertion_pct)/100*5*0.30>,
  "quality_verdict": "high (>=4.0) / medium (2.5-3.9) / low (<2.5)",
  "notes": "one sentence summary of overall quality"
}}

Respond with ONLY the JSON object, no other text.\
"""


def _avg_grounding(scores_list: list[dict]) -> dict:
    sg  = [s.get("evidence_grounding", {}).get("search_grounded_pct", 0) for s in scores_list]
    mk  = [s.get("evidence_grounding", {}).get("model_knowledge_pct", 0) for s in scores_list]
    as_ = [s.get("evidence_grounding", {}).get("assertion_pct", 0) for s in scores_list]  # trailing _ avoids shadowing built-in `assert`
    n = len(sg)
    return {"search": sum(sg)/n, "model": sum(mk)/n, "assertion": sum(as_)/n}


def _call_quality(client: genai.Client, query: str, response: str) -> dict:
    prompt = QUALITY_PROMPT.format(query=query, response=response)
    config = types.GenerateContentConfig(
        temperature=0.1,
        max_output_tokens=1024,
        response_mime_type="application/json",
        thinking_config=types.ThinkingConfig(thinking_budget=0),
    )
    text = client.models.generate_content(model=MODEL, contents=prompt, config=config).text.strip()
    return json.loads(text)


def run_quality(results_path: str = "eval/results.json",
                output_path: str = "eval/quality_results.json"):
    with open(results_path) as f:
        results = json.load(f)

    api_key = os.getenv("GEMINI_API_KEY")
    if not api_key:
        print("ERROR: GEMINI_API_KEY not set.")
        sys.exit(1)
    client = genai.Client(api_key=api_key)

    quality_results = []
    failures = []
    for i, pair in enumerate(results):
        pid = pair["id"]
        print(f"[{i+1}/{len(results)}] scoring {pid}...")

        if _has_pipeline_failure(pair):
            print(f"  SKIP — pipeline failure")
            failures.append(pair)
            continue

        try:
            score_a = _call_quality(client, pair["query_a"], pair["response_a"])
            qa = score_a.get("overall_quality", "?")
            va = score_a.get("quality_verdict", "?")
            print(f"  A: overall={qa:.2f}  verdict={va}")
        except Exception as e:
            print(f"  ERROR scoring A: {e}")
            score_a = {"error": str(e)}

        try:
            score_b = _call_quality(client, pair["query_b"], pair["response_b"])
            qb = score_b.get("overall_quality", "?")
            vb = score_b.get("quality_verdict", "?")
            print(f"  B: overall={qb:.2f}  verdict={vb}")
        except Exception as e:
            print(f"  ERROR scoring B: {e}")
            score_b = {"error": str(e)}

        quality_results.append({
            "id": pid,
            "category": pair["category"],
            "query_a": pair["query_a"],
            "query_b": pair["query_b"],
            "score_a": score_a,
            "score_b": score_b,
        })

    with open(output_path, "w") as f:
        json.dump(quality_results, f, indent=2)

    # Summary
    valid = [
        r for r in quality_results
        if "overall_quality" in r.get("score_a", {}) and "overall_quality" in r.get("score_b", {})
    ]
    if not valid:
        print("\nNo valid quality results.")
        return [], failures

    print("\n---- ARGUMENT QUALITY SCORES ----")
    print(f"{'ID':<30} {'Cat':<22} {'A score':<10} {'B score':<10} {'Gap':<8} {'A verdict':<12} {'B verdict'}")
    print("-" * 100)

    gap_pairs = []
    for r in sorted(valid, key=lambda x: abs(
        x["score_a"]["overall_quality"] - x["score_b"]["overall_quality"]
    ), reverse=True):
        qa = r["score_a"]["overall_quality"]
        qb = r["score_b"]["overall_quality"]
        gap = abs(qa - qb)
        va = r["score_a"].get("quality_verdict", "?")
        vb = r["score_b"].get("quality_verdict", "?")
        flag = " <- gap" if gap >= QUALITY_GAP_THRESHOLD else ""
        print(f"{r['id']:<30} {r['category']:<22} {qa:<10.2f} {qb:<10.2f} {gap:<8.2f} {va:<12} {vb}{flag}")
        if gap >= QUALITY_GAP_THRESHOLD:
            gap_pairs.append(r)

    all_a = [r["score_a"]["overall_quality"] for r in valid]
    all_b = [r["score_b"]["overall_quality"] for r in valid]
    all_scores = all_a + all_b

    print(f"\n{'Mean overall (A):':<30} {sum(all_a)/len(all_a):.2f}")
    print(f"{'Mean overall (B):':<30} {sum(all_b)/len(all_b):.2f}")
    print(f"{'Mean overall (all):':<30} {sum(all_scores)/len(all_scores):.2f}  (1=low, 5=high)")

    # Per-category breakdown
    by_cat: dict = defaultdict(lambda: {"a": [], "b": []})
    for r in valid:
        by_cat[r["category"]]["a"].append(r["score_a"]["overall_quality"])
        by_cat[r["category"]]["b"].append(r["score_b"]["overall_quality"])

    print("\n----- BY CATEGORY ----")
    print(f"{'Category':<22} {'Mean A':<10} {'Mean B':<10} {'Δ (A-B)'}")
    print("-" * 52)
    for cat, scores in sorted(by_cat.items()):
        mean_a = sum(scores["a"]) / len(scores["a"])
        mean_b = sum(scores["b"]) / len(scores["b"])
        delta = mean_a - mean_b
        flag = " <- A higher" if delta > 0.5 else (" <- B higher" if delta < -0.5 else "")
        print(f"{cat:<22} {mean_a:<10.2f} {mean_b:<10.2f} {delta:+.2f}{flag}")

    # Evidence grounding breakdown
    grounding_a = _avg_grounding([r["score_a"] for r in valid])
    grounding_b = _avg_grounding([r["score_b"] for r in valid])

    print("\n=== EVIDENCE GROUNDING (avg % across all responses) ===")
    print(f"{'Source':<18} {'A responses':<16} {'B responses'}")
    print("-" * 46)
    print(f"{'Search-grounded':<18} {grounding_a['search']:<16.1f} {grounding_b['search']:.1f}")
    print(f"{'Model knowledge':<18} {grounding_a['model']:<16.1f} {grounding_b['model']:.1f}")
    print(f"{'Bare assertion':<18} {grounding_a['assertion']:<16.1f} {grounding_b['assertion']:.1f}")

    # Quality gaps
    if gap_pairs:
        print(f"\n---- QUALITY GAPS ({len(gap_pairs)} pairs) ----")
        print("Quality gaps usually reflect search result asymmetry (Tavily found richer")
        print("sources for one side), not model bias. Cross-reference with judge.py results.")
        for r in gap_pairs:
            qa = r["score_a"]["overall_quality"]
            qb = r["score_b"]["overall_quality"]
            higher = "A" if qa > qb else "B"
            print(f"  {r['id']} ({r['category']}) — A={qa:.2f} B={qb:.2f}  [{higher} higher]")
    else:
        print("\nNo pairs exceed the quality gap threshold.")

    # Pipeline reliability
    _print_pipeline_reliability(failures, len(valid) + len(failures))

    print(f"\nFull results saved to {output_path}")

    return valid, failures


if __name__ == "__main__":
    run_quality()
