requirements.txt
gradio
python-dotenv
google-genai
tavily-python
sentence-transformers
numpy
